+++ db-async-common/src/main/scala/com/github/mauricio/async/db/Configuration.scala
@@ -17,11 +17,10 @@
 package com.github.mauricio.async.db
 
 import java.nio.charset.Charset
-import scala.Predef._
-import scala.{None, Option, Int}
+
+import io.netty.buffer.{ByteBufAllocator, PooledByteBufAllocator}
 import io.netty.util.CharsetUtil
-import io.netty.buffer.AbstractByteBufAllocator
-import io.netty.buffer.PooledByteBufAllocator
+
 import scala.concurrent.duration._
 
 object Configuration {
@@ -44,6 +43,11 @@ object Configuration {
  *                           OOM or eternal loop attacks the client could have, defaults to 16 MB. You can set this
  *                           to any value you would like but again, make sure you know what you are doing if you do
  *                           change it.
+ * @param allocator the netty buffer allocator to be used
+ * @param connectTimeout the timeout for connecting to servers
+ * @param testTimeout the timeout for connection tests performed by pools
+ * @param queryTimeout the optional query timeout
+ *
  */
 
 case class Configuration(username: String,
@@ -53,7 +57,7 @@ case class Configuration(username: String,
                          database: Option[String] = None,
                          charset: Charset = Configuration.DefaultCharset,
                          maximumMessageSize: Int = 16777216,
-                         allocator: AbstractByteBufAllocator = PooledByteBufAllocator.DEFAULT,
+                         allocator: ByteBufAllocator = PooledByteBufAllocator.DEFAULT,
                          connectTimeout: Duration = 5.seconds,
-                         testTimeout: Duration = 5.seconds
-                          )
+                         testTimeout: Duration = 5.seconds,
+                         queryTimeout: Option[Duration] = None)