+++ mysql-async/src/test/scala/com/github/mauricio/async/db/mysql/BinaryColumnsSpec.scala
@@ -2,6 +2,8 @@ package com.github.mauricio.async.db.mysql
 
 import org.specs2.mutable.Specification
 import java.util.UUID
+import java.nio.ByteBuffer
+import io.netty.buffer.Unpooled
 import io.netty.util.CharsetUtil
 import com.github.mauricio.async.db.RowData
 
@@ -94,6 +96,52 @@ class BinaryColumnsSpec extends Specification with ConnectionHelper {
           row("varbinary_column") === bytes
       }
 
+    }
+
+    "support BLOB type" in {
+
+      val bytes = (1 to 10).map(_.toByte).toArray
+
+      testBlob(bytes)
+
+    }
+
+    "support BLOB type with large values" in {
+
+      val bytes = (1 to 2100).map(_.toByte).toArray
+
+      testBlob(bytes)
+
+    }
+
+  }
+
+  def testBlob(bytes: Array[Byte]) = {
+    val create =
+      """CREATE TEMPORARY TABLE POSTS (
+        | id INT NOT NULL,
+        | blob_column BLOB,
+        | primary key (id))
+      """.stripMargin
+
+    val insert = "INSERT INTO POSTS (id,blob_column) VALUES (?,?)"
+    val select = "SELECT id,blob_column FROM POSTS ORDER BY id"
+
+    withConnection {
+      connection =>
+        executeQuery(connection, create)
+        executePreparedStatement(connection, insert, 1, Some(bytes))
+        executePreparedStatement(connection, insert, 2, ByteBuffer.wrap(bytes))
+        executePreparedStatement(connection, insert, 3, Unpooled.wrappedBuffer(bytes))
+
+        val Some(rows) = executeQuery(connection, select).rows
+        rows(0)("id") === 1
+        rows(0)("blob_column") === bytes
+        rows(1)("id") === 2
+        rows(1)("blob_column") === bytes
+        rows(2)("id") === 3
+        rows(2)("blob_column") === bytes
+        rows.size === 3
     }
 
   }