+++ postgresql-async/src/main/scala/com/github/mauricio/async/db/postgresql/PostgreSQLConnection.scala
@@ -20,6 +20,7 @@ import com.github.mauricio.async.db.QueryResult
 import com.github.mauricio.async.db.column.{ColumnEncoderRegistry, ColumnDecoderRegistry}
 import com.github.mauricio.async.db.exceptions.{InsufficientParametersException, ConnectionStillRunningQueryException}
 import com.github.mauricio.async.db.general.MutableResultSet
+import com.github.mauricio.async.db.pool.TimeoutScheduler
 import com.github.mauricio.async.db.postgresql.codec.{PostgreSQLConnectionDelegate, PostgreSQLConnectionHandler}
 import com.github.mauricio.async.db.postgresql.column.{PostgreSQLColumnDecoderRegistry, PostgreSQLColumnEncoderRegistry}
 import com.github.mauricio.async.db.postgresql.exceptions._
@@ -45,10 +46,11 @@ class PostgreSQLConnection
   encoderRegistry: ColumnEncoderRegistry = PostgreSQLColumnEncoderRegistry.Instance,
   decoderRegistry: ColumnDecoderRegistry = PostgreSQLColumnDecoderRegistry.Instance,
   group : EventLoopGroup = NettyUtils.DefaultEventLoopGroup,
-  executionContext : ExecutionContext = ExecutorServiceUtils.CachedExecutionContext
+  implicit val executionContext : ExecutionContext = ExecutorServiceUtils.CachedExecutionContext
   )
   extends PostgreSQLConnectionDelegate
-  with Connection {
+  with Connection
+  with  TimeoutScheduler {
 
   import PostgreSQLConnection._
 
@@ -63,7 +65,6 @@ class PostgreSQLConnection
 
   private final val currentCount = Counter.incrementAndGet()
   private final val preparedStatementsCounter = new AtomicInteger()
-  private final implicit val internalExecutionContext = executionContext
 
   private val parameterStatus = new scala.collection.mutable.HashMap[String, String]()
   private val parsedStatements = new scala.collection.mutable.HashMap[String, PreparedStatementHolder]()
@@ -80,6 +81,7 @@ class PostgreSQLConnection
   
   private var queryResult: Option[QueryResult] = None
 
+  override def eventLoopGroup : EventLoopGroup = group
   def isReadyForQuery: Boolean = this.queryPromise.isEmpty
 
   def connect: Future[Connection] = {
@@ -91,6 +93,7 @@ class PostgreSQLConnection
   }
 
   override def disconnect: Future[Connection] = this.connectionHandler.disconnect.map( c => this )
+  override def onTimeout = disconnect
 
   override def isConnected: Boolean = this.connectionHandler.isConnected
 
@@ -103,7 +106,7 @@ class PostgreSQLConnection
     this.setQueryPromise(promise)
 
     write(new QueryMessage(query))
-
+    addTimeout(promise,configuration.queryTimeout)
     promise.future
   }
 
@@ -130,7 +133,7 @@ class PostgreSQLConnection
         holder.prepared = true
         new PreparedStatementOpeningMessage(holder.statementId, holder.realQuery, values, this.encoderRegistry)
       })
-
+    addTimeout(promise,configuration.queryTimeout)
     promise.future
   }
 
@@ -303,6 +306,7 @@ class PostgreSQLConnection
 
   private def succeedQueryPromise(result: QueryResult) {
     this.queryResult = None
+    this.currentQuery = None
     this.clearQueryPromise.foreach {
       _.success(result)
     }