+++ postgresql-async/src/test/scala/com/github/mauricio/async/db/postgresql/pool/ConnectionPoolSpec.scala
@@ -16,11 +16,20 @@
 
 package com.github.mauricio.async.db.postgresql.pool
 
+import java.util.UUID
+
 import com.github.mauricio.async.db.pool.{ConnectionPool, PoolConfiguration}
+import com.github.mauricio.async.db.postgresql.exceptions.GenericDatabaseException
 import com.github.mauricio.async.db.postgresql.{PostgreSQLConnection, DatabaseTestHelper}
 import org.specs2.mutable.Specification
 
+object ConnectionPoolSpec {
+  val Insert = "insert into transaction_test (id) values (?)"
+}
+
 class ConnectionPoolSpec extends Specification with DatabaseTestHelper {
+
+  import ConnectionPoolSpec.Insert
 
   "pool" should {
 
@@ -49,6 +58,29 @@ class ConnectionPoolSpec extends Specification with DatabaseTestHelper {
         pool =>
           await(pool.connect) === pool
       }
+    }
+
+    "runs commands for a transaction in a single connection" in {
+
+      val id = UUID.randomUUID().toString
+
+      withPool {
+        pool =>
+          val operations = pool.inTransaction {
+            connection =>
+              connection.sendPreparedStatement(Insert, List(id)).flatMap {
+                result =>
+                  connection.sendPreparedStatement(Insert, List(id)).map {
+                    failure =>
+                      List(result, failure)
+                  }
+              }
+          }
+
+          await(operations) must throwA[GenericDatabaseException]
+
+      }
+
     }
 
   }