+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 **Table of Contents**
 
 - [Changelog](#changelog)
+	- [0.2.17 - in progresss](#0217---in-progresss)
+	- [0.2.16 - 2015-01-04](#0216---2015-01-04)
+	- [0.2.15 - 2014-09-12](#0215---2014-09-12)
 	- [0.2.14 - 2014-08-30](#0214---2014-08-30)
 	- [0.2.13 - 2014-04-07](#0213---2014-04-07)
 	- [0.2.12 - 2014-01-11](#0212---2014-01-11)
@@ -22,7 +25,25 @@
 
 # Changelog
 
-## 0.2.15 - still in progress
+## 0.2.18 - 2015-08-08
+
+* Timeouts implemented queries for MySQL and PostgreSQL - @lifey - #147
+
+## 0.2.17 - 2015-07-13
+
+* Fixed pool leak issue - @haski
+* Fixed date time formatting issue - #142
+
+## 0.2.16 - 2015-01-04
+
+* Add support to byte arrays for PostgreSQL 8 and older - @SattaiLanfear - #21;
+* Make sure connections are returned to the pool before the result is returned to the user - @haski - #119;
+* Support to `SEND_LONG_DATA` to MySQL - @mst-appear - #115;
+* Support for `ByteBuffer` and `ByteBuf` for binary data - @mst-appear - #113 #112;
+* Fixed encoding backslashes in PostgreSQL arrays - @dylex - #110;
+* Included `escape` encoding method for bytes in PostgreSQL - @SattaiLanfear - #107;
+
+## 0.2.15 - 2014-09-12
 
 * Fixes issue where PostgreSQL decoders fail to produce a NULL value if the null is wrapped by a `Some` instance - #99;
 * Fixes issue where the 253 case of length encoded fields on MySQL produce a wrong value;